requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Webhooks
svix==1.21.0

# Weaviate
weaviate-client==3.25.3

# LLM (Optional - for AI chat features)
openai>=1.0.0

